moved everything in src/main/scala so that directory matches package
