+++ src/main/scala/freechips/rocketchip/amba/axi4stream/AXI4StreamWidthAdapter.scala
