+++ src/main/scala/freechips/rocketchip/amba/axi4stream/Buffer.scala
