+++ src/main/scala/freechips/rocketchip/amba/axi4stream/SimpleSplitter.scala
