+++ src/main/scala/freechips/rocketchip/jtag2mm/JtagFuzzer.scala
@@ -3,9 +3,9 @@
 package freechips.rocketchip.jtag2mm
 
 import chisel3._
+import chisel3.experimental._
 import chisel3.util._
 import chisel3.util.random.LFSR
-import chisel3.experimental._
 
 
 class InvertedJtagIO extends Bundle {
