+++ src/main/scala/freechips/rocketchip/jtag2mm/JtagToMaster.scala
@@ -3,14 +3,12 @@
 package freechips.rocketchip.jtag2mm
 
 import chisel3._
-import chisel3.util._
 import chisel3.experimental._
+import chisel3.util._
+import freechips.rocketchip.TLToBundleBridge
 //import chisel3.experimental.{withClockAndReset}
 
 import freechips.rocketchip.amba.axi4._
-import freechips.rocketchip.amba.axi4stream._
-import freechips.rocketchip.config._
-import freechips.rocketchip.diplomacy._
 import freechips.rocketchip.tilelink._
