+++ src/test/scala/freechips/rocketchip/amba/axi4stream/AXI4StreamSpec.scala
@@ -4,6 +4,7 @@ import breeze.stats.distributions.Uniform
 import chisel3._
 import chisel3.iotesters.PeekPokeTester
 import dsptools.tester.MemMasterModel
+import freechips.rocketchip.{BundleBridgeToTL, TLMasterModel}
 import freechips.rocketchip.amba.axi4._
 import freechips.rocketchip.amba.axi4stream
 import freechips.rocketchip.config.Parameters
